extracted ui system into reusable library crate
- created cross-magic-ui library with proper module structure
- moved core ui components (UIRect, Widget trait, UIManager) to core.rs
- organized layout system (UIContainer, Layout types) in layout.rs
- consolidated built-in widgets (Button, Text, Panel) in widgets.rs
- added comprehensive documentation and examples
- updated main project to use library dependency
- removed old ui files from main project using git rm
- maintained full backward compatibility with existing functionality
