+++ Cargo.toml
@@ -10,3 +10,4 @@ rand = "0.9.1"
 reqwest = { version = "0.12.23", features = [ "blocking" ] }
 serde = { version = "1.0.219", features = [ "derive" ] }
 serde_json = "1.0.140"
+cross-magic-ui = { path = "ui-lib" }
