+++ ui-lib/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "cross-magic-ui"
+version = "0.1.0"
+edition = "2021"
+description = "A widget-based UI library for macroquad games"
+license = "MIT OR Apache-2.0"
+authors = ["Generated with Claude Code"]
+
+[dependencies]
+macroquad = "0.4.14"
+
+[lib]
+name = "cross_magic_ui"
+path = "src/lib.rs"
